+++ source/contracts/libraries/math/SafeMathUint256.sol
@@ -57,13 +57,4 @@ library SafeMathUint256 {
   function isMultipleOf(uint256 a, uint256 b) internal pure returns (bool) {
     return a % b == 0;
   }
-
-  // Float [fixed point] Operations
-  function fxpMul(uint256 a, uint256 b, uint256 base) internal pure returns (uint256) {
-    return div(mul(a, b), base);
-  }
-
-  function fxpDiv(uint256 a, uint256 b, uint256 base) internal pure returns (uint256) {
-    return div(mul(a, base), b);
-  }
 }