+++ tests/solidity_test_helpers/SafeMathInt256Tester.sol
@@ -37,12 +37,4 @@ contract SafeMathInt256Tester {
   function getInt256Max() public view returns (int256) {
     return SafeMathInt256.getInt256Max();
   }
-
-  function fxpMul(int256 _a, int256 _b, int256 _base) public view returns (int256) {
-    return SafeMathInt256.fxpMul(_a, _b, _base);
-  }
-
-  function fxpDiv(int256 _a, int256 _b, int256 _base) public view returns (int256) {
-    return SafeMathInt256.fxpDiv(_a, _b, _base);
-  }
 }